Minor copy change for improved clarity
Recently several of my colleagues and I found the phrase "This
stops the Observable" to be misleading in the context of multiple
observers of an observable. While the observable stops with
respect to that observer, it may still call onNext for other
observers and so the Observable itself is arguably not stopped.
This confusion gave us the mistaken impression that an observable
needs to be recreated after termination, when in fact
resubscription can be sufficient. We didn't discover our mistake
until we encountered the retry operator, which would have been
useless under our original interpretation of the onError and
onCompleted methods.